fix: wave-2 review findings across replication, pack and admission (#9)
Snapshot chunking had no way to be reassembled. splitSnapshot cut one
SnapshotBatch into N datagrams that all carried the same server_tick and
nothing else, so neither receiver could tell a fragment from a whole
world: the Go client returned the first chunk, the C# timeline replaced
the tick with the last one, and the zone loop despawned every entity that
landed in a sibling chunk. At 28 entities -- one more than InstLeague1
spawns -- every snapshot split, so half the zone was destroyed and
recreated each frame on the datagram path.

SnapshotBatch now carries chunk_index and chunk_count. A whole batch says
chunk_count 1, including on the reliable fallback, so a receiver applies
one completeness rule rather than inferring it from the carrier, and
Client.ReadSnapshot holds a tick until every chunk of it has arrived. The
split reserves room for the two fields before measuring, and an entity
whose own envelope will not fit a datagram is dropped and counted instead
of emitted: content_id, name_key and faction are unbounded pack strings,
and emitting it failed MarshalUnreliable, which ended the session for
every player in the zone.

The rest, in the order the review numbered them:

- pack.openTable computed its region lengths in uint32. row_count
  0x40000000 wrapped 12*row_count to 0 and 4*(row_count+1) to 4, both of
  which fit a 48-byte file, and the row-index walk then read off the end
  and panicked instead of returning ErrMalformedTable. The Rust verifier
  uses checked arithmetic, so the two readers accepted different files.
- The shard never set session.Server.PackID, so ADR 0027's content check
  was unreachable and any client with any pack was admitted -- while a
  client that did name its pack could never connect, because ServerHello
  carried an empty one. The shard now states the digest it loaded, and
  content.allow_unverified_pack decides whether a client that names none
  is taken (protocol/session.md rule 5.1.4).
- A refusal raised inside a session was written and then the connection
  was closed outright, and a QUIC CONNECTION_CLOSE does not wait for
  stream data: the peer saw an opaque abort instead of the reason. Both
  refusal paths now half-close, drain, then close. Every refusal test
  used net.Pipe, which cannot show this, so the new ones use real QUIC.
- Pre-admission reads observed neither context nor deadline. A peer that
  finished the QUIC handshake and never wrote a ClientHello pinned a
  goroutine for the life of the process and survived SIGTERM. The
  watchdog is now armed before the first read and admission has its own
  deadline. Worse, quicListener.Accept accepted the session stream on
  the accept loop's own goroutine, and a QUIC stream is invisible until
  its opener writes -- so one silent peer blocked admission for everyone.
- pack.Load followed any path a manifest named, turning a pack directory
  into a size-and-hash oracle for arbitrary files, and ignored .sptbl
  files the manifest did not list, so two different directories shared a
  pack_id. Both are now refused, matching the Rust verifier.
- The in-memory store claimed parity with Postgres while enforcing
  neither the character-name length check, the account_id foreign key,
  nor the level, experience and save_seq checks. constraints.go states
  them once; Postgres maps 23514 and 23503 onto the same sentinel.
- ALTER DEFAULT PRIVILEGES so a later migration's tables are covered by
  the ownership boundary, and a role drop that reports and survives a
  cluster-global dependency instead of failing the whole rollback.
- runAndRecover kept the panic message and threw away the stack.

Refs: docs/specs/protocol/session.md 5.1.4, 5.5.7; ADR 0027, 0029, 0031

+++ internal/pack/table.go
@@ -69,20 +69,26 @@ func openTable(name string, payload []byte) (*table, error) {
 	}
 	rowDataBytes := binary.LittleEndian.Uint32(payload[28:])
 
+	// Every region length is computed in uint64. In uint32 the products wrap:
+	// row_count 0x40000000 makes the key index 12*row_count = 0 bytes long and
+	// the row index 4*(row_count+1) = 4, both of which fit a 48-byte file, and
+	// the row-index walk below then runs off the end of the payload. The Rust
+	// verifier uses checked arithmetic for the same reason, and the two readers
+	// must accept exactly the same set of files.
 	regions := []struct {
 		name   string
 		offset uint32
-		length uint32
+		length uint64
 	}{
-		{"key index", loaded.keyIndexOffset, tableKeyEntrySize * loaded.rowCount},
-		{"row index", loaded.rowIndexOffset, 4 * (loaded.rowCount + 1)},
-		{"row data", loaded.rowDataOffset, rowDataBytes},
+		{"key index", loaded.keyIndexOffset, uint64(tableKeyEntrySize) * uint64(loaded.rowCount)},
+		{"row index", loaded.rowIndexOffset, 4 * (uint64(loaded.rowCount) + 1)},
+		{"row data", loaded.rowDataOffset, uint64(rowDataBytes)},
 	}
 	for _, region := range regions {
 		if region.offset < tableHeaderBytes {
 			return fail("%s at offset %d overlaps the header", region.name, region.offset)
 		}
-		end := uint64(region.offset) + uint64(region.length)
+		end := uint64(region.offset) + region.length
 		if end > uint64(len(payload)) {
 			return fail("%s ends at %d, past the %d byte file", region.name, end, len(payload))
 		}
@@ -92,7 +98,8 @@ func openTable(name string, payload []byte) (*table, error) {
 		if first.length == 0 || second.length == 0 {
 			continue
 		}
-		if first.offset < second.offset+second.length && second.offset < first.offset+first.length {
+		if uint64(first.offset) < uint64(second.offset)+second.length &&
+			uint64(second.offset) < uint64(first.offset)+first.length {
 			return fail("%s and %s regions overlap", first.name, second.name)
 		}
 	}
